### 2025-06-20 62dd2d8
fix missing/invalid DOIs in paper.bib
