+++ paper.bib
@@ -86,6 +86,7 @@
   address = {Vienna, Austria},
   year = {2025},
   url = {https://www.R-project.org/}
+  doi = {10.32614/r.manuals}
 }  
   
 @Book{ggplot2,
@@ -200,7 +201,7 @@
   number = {4},
   pages = {450-464},
   keywords = {Conservation, dams, fishways, habitat fragmentation, migration, passage efficiency},
-  doi = {doi.org/10.1111/j.1467-2979.2011.00445.x},
+  doi = {10.1111/j.1467-2979.2011.00445.x},
   url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1467-2979.2011.00445.x},
   eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1467-2979.2011.00445.x},
   year = {2012}
@@ -246,10 +247,10 @@
 }
 
 @Manual{anadrofish,
-  title = {anadrofish: Anadromous Fish Population Responses to Habitat Changes},
+  title = {{anadrofish}: Anadromous Fish Population Responses to Habitat Changes},
   author = {{Stich}, D.~S. and {Hardesty}, J.~D. and {Jordan}, N.~T. and {Roy}, S.~G. and {Sheehan}, T.~F. and {Snyder}, S.~D. and {Zydlewski}, J.~D.},
   note = {R package version 2.3.0},
   url = {https://danstich.github.io/anadrofish/},
   year = {2025},
-  doi = {https://doi.org/10.5281/zenodo.14285894}
+  doi = {10.5281/zenodo.14285894}
 }
