### 2025-06-20 #8427 @danStich
Hi @kthyng, this sounds good, thank you! I fixed the missing and invalid DOIs and will start looking for potential reviewers using the resources you linked.
