+++ paper.bib
@@ -85,7 +85,7 @@
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
   year = {2025},
-  url = {https://www.R-project.org/}
+  url = {https://www.R-project.org/},
   doi = {10.32614/r.manuals}
 }  
   
