### 2025-06-25 #8427 @danStich
That's great to hear! I apologize for the delay. We've finally had good weather here, so I have been in the field the past three days. I will finish the list of reviewers by the end of the day tomorrow.

Thanks very much,

Dan

----
Daniel S. Stich (he/him)
Associate Professor, Biology Department and
Secretary-Treasurer, NY Chapter American Fisheries Society
113A Perna Science, SUNY Oneonta, NY 13820
Office: 607-436-3734
Cell: 518-860-4107
Email: ***@***.***
Website: danstich.github.io/stich<https://danstich.github.io/stich/index.html>
