### 2025-06-26 #8427 @danStich
Hi @ethanwhite, I've put together a list of 5 potential reviewers from the list of people who have already agreed to review for the journal based on their profile "expertise" and "domain" descriptions. I also tried to provide a brief rationale statement for why they may be suitable. Here are their GitHub handles in alphabetical order:

* <b>egouldo</b> (Earth Sciences and Ecology): specific expertise in ecological modeling and metascience as well as decision analysis

* <b>hvoltbb</b> (Biomedical Engineering, Biosciences, Chemistry, and Materials Earth Sciences and Ecology): specific experties in fishery and conservation biology

* <b>ifoxfoot</b> (Statistics, GIS, ecology, Conservation): experience programming agent-based simulation R packages used in conservation/ecology

* <b>k-doering-NOAA</b> (Earth Sciences and Ecology): specific expertise in fisheries ecology

* <b>rmk118</b> (Fisheries and aquaculture, ecology, dynamic simulation models (e.g., Dynamic Energy Budget), data visualization, basic stats/modeling (incl. geospatial + time series): experience with dynamic simulation models and specific expertise in fisheries; previously reviewed a JOSS paper for a similar but narrower, single-species tool (dia R package)
