### 2025-06-27 #8427 @ethanwhite
👋 @ifoxfoot @rmk118 @egouldo - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
