### 2025-07-08 #8427 @k-doering-NOAA
@ethanwhite I would be happy to review this submission by July 23, if that sounds ok.
