### 2025-07-10 #8564 @editorialbot
[REVIEW]: anadrofish: R package for modeling anadromous fish population responses to dams
**Submitting author:** @danStich (<a href="http://orcid.org/0000-0002-8946-1115">Daniel Stich</a>)
**Repository:** https://github.com/danStich/anadrofish
**Branch with paper.md** (empty if default branch): master
**Version:** v2.3.1
