### 2025-07-10 danStich/anadrofish#5 @rmk118
[JOSS] Performance claim
The JOSS checklist includes the item "If there are any performance claims of the software, have they been confirmed?"
I am currently unable to complete this item because the paper says anadrofish is “is faster than other empirical life-history models (e.g., shadia),” but I did not see confirmation of this performance claim.
