### 2025-07-10 danStich/anadrofish#6 @rmk118
Images not rendering on Github Pages site
The conceptual diagram and plots resulting from the example usage do not appear to be rendering properly on the main Github Pages site: https://danstich.github.io/anadrofish/, I am just seeing the placeholder missing image icon
