### 2025-07-10 danStich/anadrofish#7 @rmk118
[JOSS] Community Guidelines
JOSS Checklist Item: Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support

I would recommend removing the links to the FAQ and Discussions pages in the Contributions.md file unless these are enabled, as they currently bring the user to the repo home page and a 404 page, respectively.

Additionally, the Contributing page says that coding contributions should adhere to the tidyverse style guide, but there are a few deviations from tidyverse style in the existing code (e.g., missing spaces in `retlist <- list(res=res)` and `juvenile_s <- exp(-Zd*30)`). A formatter like [Air](https://posit-dev.github.io/air/formatter.html) or [lintr](https://lintr.r-lib.org/) might be helpful to ensure consistent code styling.
