### 2025-07-10 danStich/anadrofish#8 @rmk118
[JOSS] Miscellaneous
Hi Dan,

It’s great to review another one of your packages and to see how this one expands upon and combines the functionality of your previous work! Here are some additional comments that felt  too minor to merit individual issues, but I’m happy to split them up into separate issues if that would be better for your workflow.

1. The LICENSE file is a .md file, but the checklist wording suggests it may need to be a .txt? I’m not sure how strict they are about this.

2. Notes related to the examples:
- “Approximately 1-6 million spawning alewife have passed upstream of Benton Falls annually since then.” Potentially clarify the location of Benton Falls in relation to the removed or modified dams.
- What does the dashed horizontal line at 1.3 million spawners represent in the figure for the American shad Connecticut River example?

3. Notes related to the documentation of sim_pop_new.R:
- I would recommend changing the documentation that currently states that the natural mortality argument of `sim_pop`, unless specified otherwise, uses the `mortality` dataset that contains sex-specific instantaneous mortality of American shad.
         - For instance, the first example is simulating alewife in the Sebasticook but the `nM` argument is not specified, so I thought it was pulling from the American shad mortality dataset. Looking at the source code, my understanding of what is actually happening (in the alewife/Sebasticook example) is that `sim_pop` calls `make_mortality()` for each sex with `river = Sebasticook`, `species = ALE`, and `custom_habitat = sebasticook_habitat`. `make_mortality()` then calls `mortality_rh()` to get sex-specific mortality rates for alewife in the NNE region.
        - I would make analogous changes to `sim_pop()` documentation descriptions of the spawnRecruit and max_age parameters to clarify that they do not use American shad-specific datasets when modeling the other species.
- Change the `s_juvenile` description to clarify that only AMS survival from hatch to outmigrant is calculated using crecco_1983, for BBH it is based on Overton et al. (2012) and for ALE it uses Hook et al. (2007).
- Clarify in the `river` parameter description that the user should put the custom river name here if specifying a `custom_habitat`, as the current description implies that only rivers from the `get_rivers()` function are accepted.
- Consider clarifying that the downstream argument of `sim_pop` specifically refers to only adult fish.

4. Other
- Remove the extra #' before `@references` in crecco_1983.R to ensure proper rendering in the documentation.
- There is a minor spelling error in mortality.R: change instaneous to instantaneous.
- I am curious why you created the `make_spawners()` and `make_recruits()` functions that do not appear to offer any functionality beyond R’s default multiplication.
