### 2025-07-10 #8564 @rmk118
I have started working on my checklist and opened several issues in the target repository: https://github.com/danStich/anadrofish/issues/5, https://github.com/danStich/anadrofish/issues/6, https://github.com/danStich/anadrofish/issues/7, https://github.com/danStich/anadrofish/issues/8
