### 2025-07-10 bd540e0
pkgdown test for issue 6
