+++ pkgdown.yml
@@ -2,7 +2,7 @@ pandoc: 3.1.11
 pkgdown: 2.1.3
 pkgdown_sha: ~
 articles: {}
-last_built: 2025-07-10T21:42Z
+last_built: 2025-07-10T21:59Z
 urls:
   reference: https://danstich.github.io/anadrofish/reference
   article: https://danstich.github.io/anadrofish/articles
