### 2025-07-10 7dd9c72
pkgdown test for issue 6
