### 2025-07-10 d4c24bb
move images to man/figures
