### 2025-07-10 3f76d8e
revert image location, test raw paths
