### 2025-07-10 c17fb1b
move images to man/figures for #6
