### 2025-07-10 danStich/anadrofish#6 @danStich
Was able to resolve this by moving image files to man/figures as suggested here: https://github.com/r-lib/pkgdown/issues/133
