### 2025-07-10 e7dc2b2
run styler on pkg and rebuild docs
