+++ tests/testthat/test-make_postspawn.R
@@ -1,6 +1,4 @@
 test_that("Returns numeric of length 1", {
-  
   expect_length(make_postspawn(river = "Hudson", species = "AMS"), 1)
-  expect_type(make_postspawn(river = "Hudson", species = "AMS"), 'double')
-  
+  expect_type(make_postspawn(river = "Hudson", species = "AMS"), "double")
 })
