+++ R/vbgf_bbh.R
@@ -1,10 +1,10 @@
-#' @title Sex-specific von Bertalanffy growth parameters for 
+#' @title Sex-specific von Bertalanffy growth parameters for
 #' blueback herring
-#' 
+#'
 #' @description A dataset containing sex-specific estimates of von Bertalanffy
-#' growth parameters for blueback herring from the 2024 ASMFC river herring 
+#' growth parameters for blueback herring from the 2024 ASMFC river herring
 #' benchmark stock assessment (ASMFC 2024).
-#' 
+#'
 #' @format A data frame with 816,000 observations of 5 variables:
 #' \describe{
 #'     \item{\code{Region}}{Genetic reporting group}
@@ -15,7 +15,7 @@
 #' }
 #'
 #' @references Atlantic States Marine Fisheries Commission. 2024. River herring
-#' benchmark stock assessment and peer-review report. ASMFC, Arlington, VA. 
+#' benchmark stock assessment and peer-review report. ASMFC, Arlington, VA.
 #' URL: https://asmfc.org/uploads/file/66f59e40RiverHerringAssessment_PeerReviewReport_2024.pdf
 #'
 #' @source Atlantic States Marine Fisheries Commission
