+++ inst/examples/upper95_ex.R
@@ -1,8 +1,8 @@
 # Example usage
 \dontrun{
   
- x <- rnorm(1e4, 0, 1)
+x <- rnorm(1e4, 0, 1)
  
- upper95(x)
+upper95(x)
   
 }
