+++ inst/examples/simpop_ex.R
@@ -1,5 +1,5 @@
 # Example usage
-# \dontrun{
+\dontrun{
 
 # Example 1. Simulate a single population one time for 50 years ----------------
 library(anadrofish)
@@ -263,4 +263,4 @@ ggplot(coastal, aes(x = scenario, y = fit)) +
   xlab("Scenario") +
   ylab("Number of spawning adults")
 
-# }
+}
