### 2025-07-11 7407adf
rebuild doc for sim_pop()
