### 2025-07-11 danStich/anadrofish#8 @danStich
Thanks so much for your thoughtful comments and suggestions. This format works for me. I went ahead and made sub-issues so I can keep track of them how you organized your comments. I'll close the sub-issues as I work through them, but please let me know if any of your suggestions are not adequately addressed.
