### 2025-07-11 db628f2
code styling and updates in README.md
