+++ README.md
@@ -61,7 +61,7 @@ To install `anadrofish`, you will need to have `devtools` installed ahead of tim
 
 ## Examples
 ### Running one scenario for one river many times in parallel
-This example uses alewife in the Sebasticook River, ME, USA [Wipplehauser 2021](https://onlinelibrary.wiley.com/doi/full/10.1002/tafs.10292) to understand baseline population predictions following the removal of Edwards Dam in 1999, removal of Fort Halifax Dam in 2008, and installation of a fish lift at the next dam in the Sebasticook River. Approximately 1-6 million spawning alewife have passed upstream of Benton Falls annually since then.
+This example uses alewife in the Sebasticook River, ME, USA [Wipplehauser 2021](https://onlinelibrary.wiley.com/doi/full/10.1002/tafs.10292) to understand baseline population predictions following the removal of Edwards Dam in 1999, removal of Fort Halifax Dam in 2008, and installation of a fish lift at the next dam in the Sebasticook River. Approximately 1-6 million spawning alewife have passed upstream of Benton Falls (5.3 miles from mouth of Sebasticook River) annually since then.
 
 In this example, we return output from `sim_pop()` using `years = "all"` to also demonstrate the number of years needed for a stable population estimate under this scenario. If we wanted to model changes over time, we would run different scenarios for each year and return only the final year of simulations using `years = "last"`. In theory, this approach could also be used to identify stable population sizes for seeding the initial population in temporal studies, but this has not been validated.
 
@@ -86,6 +86,7 @@ ncpus <- detectCores() - 1
 sfInit(parallel = TRUE, cpus = ncpus, type = "SOCK")
 
 # Read in the Sebasticook River data ----
+# This data set is in the inst/data folder on the GitHub repo
 sebasticook_habitat_data <- read_csv("https://raw.githubusercontent.com/danStich/anadrofish/refs/heads/master/inst/data/sebasticook_habitat.csv")
 
 sebasticook_habitat_data <-
@@ -96,7 +97,7 @@ sebasticook_habitat_data <-
   )
 
 # Make a custom habitat data set for the Sebasticook River, ME
-# since it is only included as part of the larger Kennebec River 
+# since it is only included as part of the larger Kennebec River
 # watershed for river herring
 sebasticook_habitat <- data.frame(
   river = "Sebasticook",
@@ -110,31 +111,32 @@ sebasticook_habitat <- data.frame(
 )
 
 # Wrapper function ----
-sim <- function(x){
-
-# . Call simulation ----
+sim <- function(x) {
+  # . Call simulation ----
   res <- sim_pop(
     species = "ALE",
     river = "Sebasticook",
-    nyears = 50,    
+    nyears = 50,
     n_init = round(runif(1, 1e6, 22e7)),
     sr = 0.5,
     b = 0.05,
-    upstream = 1, #runif(1, 0.95, 1.00),
-    downstream = 1, #runif(1, 0.95, 1.00),
-    downstream_j = 1, #runif(1, 0.95, 1.00),
-    output_years = 'all',
+    upstream = 1, # runif(1, 0.95, 1.00), # Could draw from a uniform instead
+    downstream = 1, # runif(1, 0.95, 1.00),
+    downstream_j = 1, # runif(1, 0.95, 1.00),
+    output_years = "all",
     age_structured_output = FALSE,
     sex_specific = TRUE,
-    custom_habitat = sebasticook_habitat)
[…]
