+++ man/crecco_1983.Rd
@@ -14,11 +14,6 @@ A data frame with 24 observations of 6 variables:
     \item{\code{Si}}{Interval survival}
     \item{\code{Sc}}{Cumulative survival}
 }
-
-#' @references Crecco, V., T. Savoy, and L. Gunn. 1983. Daily mortality rates
-of larval and juvenile American shad (*Alosa sapidissima*) in the Connecticut
-River with changes in year-class strength. Canadian Journal of
-Fisheries and Aquatic Sciences 40:1719–1728.
 }
 \usage{
 crecco_1983
@@ -27,5 +22,11 @@ crecco_1983
 A built-in dataset containing estimated daily
 mortality rates for larval and juvenile American shad in the
 Connecticut River, USA 1979-1982 (Crecco et al. 1983).
+}
+\references{
+Crecco, V., T. Savoy, and L. Gunn. 1983. Daily mortality rates
+of larval and juvenile American shad (*Alosa sapidissima*) in the Connecticut
+River with changes in year-class strength. Canadian Journal of
+Fisheries and Aquatic Sciences 40:1719–1728.
 }
 \keyword{datasets}
