### 2025-07-11 74a143b
update README for #8 and #10
