### 2025-07-11 danStich/anadrofish#8 @danStich
Hi Ruby, I think I've addressed the suggestions you made in this issue through the sub-issues I created. Please let me know if I missed any of these!
