### 2025-07-11 95fada4
add benchmarking example for #5
