### 2025-07-11 danStich/anadrofish#5 @danStich
Added a minimal benchmarking example to the bottom of Examples section in `README.md`. It is small (10 evaluations) because the [shadia](https://github.com/danStich/shadia) R package takes around 30 seconds per model run, and the corresponding `anadrofish` models in `sim_pop()` take around 0.05 seconds so the difference is huge. I can increase this if you think it's necessary.
