### 2025-07-11 74836b7
re-code figure captions and alt-text using inline html for #10
