### 2025-07-11 f463aba
fix rng in example code and update figures for #13
