### 2025-07-14 603f972
add rlecuyer RNGstream for reproducibility #13
