### 2025-07-16 b6d8460
update repro examples for README #13
