### 2025-07-16 25ba1c5
correct ncpus for examples #13
