### 2025-07-16 #8564 @rmk118
Hi @ethanwhite, I wanted to let you know that I've completed my checklist! @danStich did a fantastic job addressing all of my comments (issues #1-13 in the [anadrofish repo](https://github.com/danStich/anadrofish))
