### 2025-07-18 #8564 @ethanwhite
Thanks very much for reporting this potential COI @k-doering-NOAA. There is an exemption in the [JOSS COI Policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) specifically for this case:

> some conflicts may be recorded and then waived, and if you think you are able to make an impartial assessment of the work, you should request that the conflict be waived. For example... if you and a submitter are both employed by the same very large organization but in different units without any knowledge of each other.

Given that, as long as you think you are able to make an impartial assessment of the work I'm happy to waive the institutional conflict.

I'm also going to ping the lead editor for this track (@kthyng ) just to have her confirm my reading of things since this is my first time waiving a conflict.
