### 2025-07-18 #8564 @kthyng
@ethanwhite @k-doering-NOAA For such a large organization as NOAA, we can think of different offices at different institutions in this case and waive the COI. Thank you for bringing it up.
