### 2025-07-21 danStich/anadrofish#14 @k-doering-NOAA
[JOSS] blocking issue: could not install shadia to check performance claim for anadrofish
Thanks for putting together such detailed examples on the readme! I was able to run them all, except the benchmarking example.

This is because I had issues with installing the shadia package. I would like to be able to reproduce the benchmarking example to verify the performance claim.

 I got the following error:

```r
Error: package or namespace load failed for 'shadia' in namespaceExport(ns, exports):
 undefined exports: fillOutputVectors_last
Error: loading failed
Execution halted
ERROR: loading failed
* removing 'C:/Users/kathryn.doering/AppData/Local/Programs/R/R-4.4.3/library/shadia'
Warning message:
In i.p(...) :
  installation of package ‘C:/Users/KATHRY~1.DOE/AppData/Local/Temp/RtmpUNyCue/file27a0d2f506/shadia_2.1.0.tar.gz’ had non-zero exit status
```

I investigated a bit and it looks like 3 packages needed to be added as dependencies to the DESCRIPTION file:

Namespace dependencies missing from DESCRIPTION Imports/Depends entries:
     'Rcpp', 'dplyr', 'magrittr'

In addition, there were still issues with undefined exports, which I am not sure exactly how to solve. I tried on both a linux VM and a windows machine with the same result, so I don’t think this is just an issue with my computer.
