### 2025-07-21 danStich/anadrofish#15 @k-doering-NOAA
[JOSS] suggestion: documentation
These are just suggestions, no need to address them to satisfy the JOSS reviewer checklist!

- It is possible to [categorize the functions in the reference on the pkgdown website](https://pkgdown.r-lib.org/articles/pkgdown.html#reference). This would make it easier for users to navigate the available functions in the package.
- A link back to the source code from the pkgdown site would be helpful.
- I noticed the write_output() function is intended for internal use, but it is exported. Consider not exporting it so that users can't access it. See [The exports section of the R packages book](https://r-pkgs.org/dependencies-in-practice.html#exports). Alternatively, there is an option for internal functions that aren't for users but are exported, where the documentation can be hidden by using the @noRd tag. See the [R packages book documentation workflow section](https://r-pkgs.org/man.html#sec-man-workflow):

 > "...you probably don’t want to document unexported functions. If you want to use roxygen comments for internal documentation, include the @noRd tag to suppress the creation of the .Rd file."
