### 2025-07-21 danStich/anadrofish#16 @k-doering-NOAA
[JOSS] suggestion: make tests more robust with snapshots
This is just a suggestion; no changes needed to complete the JOSS reviewer checklist.

While coverage is moderate (58%, according to running `covr::package_coverage()`), most of the tests are only checking that the return object is of the correct type. To make the test suite more robust, consider checking for exact values in addition to checking the type. In the case of complex objects, like dataframes, lists, or s3 objects, consider using [snapshot tests](https://testthat.r-lib.org/articles/snapshotting.html), which will help catch if the object output changes unexpectedly over time.
