### 2025-07-21 danStich/anadrofish#17 @k-doering-NOAA
[JOSS] suggestion: 1 paper edit
The paper looks great -  I only have one suggestion:

Lines 30 and 31: I’m not clear about what “traditional stock assessment tools” refers to here. I assumed that meant “integrated assessment models”,  but it seems like it would be possible to incorporate the complex life history into an integrated assessment model. Perhaps provide specific examples of the traditional stock assessment tools to clarify? (Also, my background is in marine and estuarine species, so please forgive my ignorance of what is considered “traditional” for diadromous fish!)
