### 2025-07-21 #8564 @k-doering-NOAA
I have completed my review, and opened issues on the anadrofish repository.

There is one blocking issue: https://github.com/danStich/anadrofish/issues/14

Three suggestions were also made: https://github.com/danStich/anadrofish/issues/17, https://github.com/danStich/anadrofish/issues/16, and https://github.com/danStich/anadrofish/issues/15
