### 2025-07-21 danStich/anadrofish#14 @danStich
Hi @k-doering-NOAA, thanks so much for your detailed feedback. I will start working through some of the other issues you opened, but wanted to address this one first.

I think I've resolved the install errors by **1)** updating the Imports section of the [DESCRIPTION file for `shadia`](https://github.com/danStich/shadia/blob/master/DESCRIPTION) and **2)** pushing the function definition file and supporting documentation for the `fillOutputVectors_last()` to the `shadia` repo, which I apparently never did after adding that option.

Could you please let me know if this fixed the issue when you get a chance?
