### 2025-07-21 danStich/anadrofish#14 @k-doering-NOAA
@danStich thank you, I was able to install shadia!

I'll try to reproduce the benchmarking example within the next few days at the latest!
