### 2025-07-22 #8564 @k-doering-NOAA
[The blocking issue](https://github.com/danStich/anadrofish/issues/14https://github.com/danStich/anadrofish/issues/14) was addressed, so I was able to check off all components of my review checklist. No further changes are required.

The remaining suggestions are up to the authors' discretion on whether they would like to make changes or not. Thanks @danStich for helping to get the blocking issue resolved so quickly!
