### 2025-07-22 danStich/anadrofish#14 @k-doering-NOAA
I was able to reproduce the benchmarking example, so this is all squared away.

Just noting for future that on my local Windows laptop my R session crashed when trying to run the example (perhaps a memory issue?), but the code worked fine on a Linux VM. I didn't explore it further, but wanted to note in case someone else runs into a similar issue in the future.
