### 2025-07-23 danStich/anadrofish#14 @danStich
That's great, thanks so much! Yes, I've also been experiencing what I assume are memory leaks on my windows machine. I think it is related to overflow from the Rcpp function in shadia that handles individual fish. I only noticed when testing in serial for benchmarking (I usually run in parallel). Hoping to swing back for that once anadrofish review is completed.

Dan

----
Daniel S. Stich (he/him)
Associate Professor, Biology Department and
Secretary-Treasurer, NY Chapter American Fisheries Society
113A Perna Science, SUNY Oneonta, NY 13820
Office: 607-436-3734
Cell: 518-860-4107
Email: ***@***.***
Website: danstich.github.io/stich<https://danstich.github.io/stich/index.html>
