### 2025-07-23 03b7fc7
paper edit for #17
