+++ paper.md
@@ -58,7 +58,7 @@ This draft manuscript is distributed solely for the purpose of peer review. Its 
 # Summary
 Diadromous fishes world-wide experienced precipitous declines during the 19^th^ and 20^th^ centuries due to a combination of overfishing, pollution, and freshwater habitat loss through construction of dams [@limburg2009]. Following wide-spread fishing closures and large-scale remediation of many historical pollution sources, dams in coastal rivers remain as the largest tractable impediment to population recovery for many of these species [@waldman2022]. In some cases, dams reduce access to as much as 95% of freshwater and rearing habitat [@hall2011]. These effects are especially pronounced for species that rely on long-distance migrations to spawning and rearing habitat upstream of barriers such as various alosines (herrings; e.g., American shad *Alosa sapidissima*, alewife *A. pseudoharengus*, and blueback herring *A. aestivalis* [@noonan2012]) and salmonines (trout and salmon; e.g., *Salmo* spp. [@parrish1998] and *Oncorhynchus*  spp. [@quinones2015]. 
 
-Because of their complex life-histories and use of discrete habitats within ontogenetic stages and seasons, traditional stock assessment tools often do not suffice for population assessment. This has historically necessitated species- and system-specific approaches to fisheries stock assessment and management strategy evaluation [e.g., @nieland2015;@barber:2018;@roy2018;@stich:2019]. Therefore, we created the `anadrofish` package [@anadrofish] for R [@R] to provide a generalized approach that also allows broader application to novel species, systems, and scenarios.
+Traditional stock assessment tools such as per-recruit analyses often fail to capture management complexities related to diadromous life histories such as fish passage at dams, and integrated assessment models can be difficult to parameterize for data-poor species such as herrings [@asmfc:2024]. This has resulted in the development of species- and system-specific approaches to fisheries stock assessment and management strategy evaluation [e.g., @nieland2015;@barber:2018;@roy2018;@stich:2019]. We created the `anadrofish` package [@anadrofish] for R [@R] to provide a generalized approach that also allows broader application to novel species, systems, and scenarios.
 
 
 # Statement of need
