### 2025-07-24 d627af1
edits for https://github.com/openjournals/joss-reviews/issues/8564#issuecomment-3108373510
