### 2025-07-24 1db536c
Update CITATION.cff
