### 2025-07-24 7758693
increment version history for https://github.com/openjournals/joss-reviews/issues/8564#issuecomment-3108373510
