### 2025-07-24 e2546c0
Update CITATION.cff
