### 2025-07-24 #8564 @danStich
Hi @ethanwhite, I think we have addressed all these as follows:
 - Removed USGS draft manuscript disclaimer
 - Removed "e.g."
 - Changed the manuscript title to "anadrofish: anadromous fish population responses to dams" to match the DESCRIPTION file (package title), incremented version number, and added a new release to Zenodo
 - Editorial suggestions
   - Changed as suggested (dropped "it")
   - Changed as suggested (added "the" and dropped comma before start of second clause for continuity)
