### 2025-08-01 #8564 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
